# public/api/docs/openapi.json
{
    "openapi": "3.0.2",
    "info": {
        "title": "Swagger",
        "description": "",
        "termsOfService": "",
        "contact": {
            "email": "apiteam@swagger.io"
        },
        "license": {
            "name": "",
            "url": ""
        },
        "version": "0.0.0"
    },
...
    "servers": [
        {
            "url": "http://0.0.0.0:5555/api/"
        }
    ],
...
    "security": [
        {
            "bearerAuth": []
        }
    ],
    "paths": {
...
        "/auth/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "LoginUser",
                "description": "LoginUser",
                "operationId": "LoginUser",
                "requestBody": {
                    "description": "LoginUser",
                    "content": {
                        "application/json": {
                            "schema": {
                                "example": {

                                    "email" : "admin@admin.com",
                                    "password" : "password"
                                }
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {

                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid ID supplied"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        },
...
        "/auth/token": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "tokenUser",
                "description": "tokenUser",
                "operationId": "tokenUser",
                "requestBody": {
                    "description": "tokenUser",
                    "content": {
                        "application/json": {
                            "schema": {
                                "example": {
                                    "device_name" : "APPLICATION_MOBILE",
                                    "email" : "admin@admin.com",
                                    "password" : "password"
                                }
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {

                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid ID supplied"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}